myHackademy/requirements.txt
blinker==1.4
Click==7.0
Django==2.1.5
django-cors-headers==2.4.0
django-extra-fields==1.0.0
django-filter==2.0.0
django-rest-framework-base64-fields==0.1.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
drf-base64==0.9.4
Flask==0.12.4
Flask-HTMLmin==1.3.2
Flask-Principal==0.4.0
Flask-WTF==0.14.2
htmlmin==0.1.12
itsdangerous==1.1.0
Jinja2==2.10
linecache2==1.0.0
Markdown==3.0.1
MarkupSafe==1.1.0
Pillow==5.4.1
psycopg2==2.7.6.1
PyJWT==1.7.1
python-dateutil==2.7.5
pytz==2018.9
simplejwt==2.0.1
six==1.12.0
typing==3.6.6
Werkzeug==0.14.1
WTForms==2.1
